added basic authentication

+++ src/main/java/com/springboot/web/demo/userDemo/service/user/UserServiceImpl.java
@@ -4,20 +4,32 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.springboot.web.demo.userDemo.model.User;
 import com.springboot.web.demo.userDemo.repository.UserRepository;
-import com.springboot.web.demo.userDemo.service.HelperService;
+import com.springboot.web.demo.userDemo.util.UserUtil;
 
 @Service
-public class UserServiceImpl implements UserService{
+public class UserServiceImpl implements UserService, UserDetailsService{
 	
 	@Autowired
 	private UserRepository userRepository;
+	
+	private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
 	public UserServiceImpl(UserRepository userRepository) {
+		this.bCryptPasswordEncoder = new BCryptPasswordEncoder();
 		this.userRepository = userRepository;
+	}
+	
+	@Override
+	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+		return findByUsername(username);
 	}
 	
 	@Override
@@ -38,6 +50,12 @@ public class UserServiceImpl implements UserService{
     		return savedUser.get();
     	}
     	else {
+    		
+    		String encodedPassword = bCryptPasswordEncoder
+                    .encode(user.getPassword());
+    		
+    		user.setPassword(encodedPassword);
+    		
     		return userRepository.save(user);
     	}
 	}
@@ -47,7 +65,7 @@ public class UserServiceImpl implements UserService{
 		Optional<User> savedUser = Optional.ofNullable(userRepository.findByUsername(username));
     	
     	if(savedUser.isPresent()) {
-			HelperService.updateRequired(savedUser.get(), user);
+			UserUtil.updateRequired(savedUser.get(), user);
     		userRepository.save(savedUser.get());
             return "user "+username+" has been updated";
     	}
@@ -57,7 +75,14 @@ public class UserServiceImpl implements UserService{
 	}
 	
 	@Override
-	public void deleteUser(String username) {
-		userRepository.deleteByUsername(username);
+	public String deleteUser(String username) {
+    	Optional<User> user = Optional.ofNullable(userRepository.findByUsername(username));
+    	if(user.isPresent()){
+    		userRepository.deleteByUsername(username);
+        	return "user "+ username +" has been deleted";
+    	}
+    	else
+    		return "user "+username+" does not exist. Aborting deletion";
+		
 	}
 }
